* using smart pointers to /base/samples/frame/Frame objects; * changed gain type (int to double); * converted orogen component interface values to SI units;  * clean code.

+++ tasks/MultibeamSonarTask.hpp
@@ -35,7 +35,7 @@ namespace imaging_sonar_simulation{
     protected:
 
 		virtual bool setRange(double value);
-		virtual bool setGain(int value);
+		virtual bool setGain(double value);
         virtual bool setNumber_of_beams(int value);
         virtual bool setNumber_of_bins(int value);
         virtual bool setAngular_resolution(double value);
